Draft: refactoring
