+++ .github/workflows/automation_lang_specifics/proto_mapping.py
@@ -1,0 +1,143 @@
+## Dictionary of proto API names, with empty methods array, to be filled in for later use by get_proto_apis():
+proto_map = {
+    "arm": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/arm/v1/arm_grpc.pb.go",
+        "name": "ArmServiceClient",
+        "methods": []
+    },
+    "base": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/base/v1/base_grpc.pb.go",
+        "name": "BaseServiceClient",
+        "methods": []
+    },
+    "board": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/board/v1/board_grpc.pb.go",
+        "name": "BoardServiceClient",
+        "methods": []
+    },
+    "camera": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/camera/v1/camera_grpc.pb.go",
+        "name": "CameraServiceClient",
+        "methods": []
+    },
+    "encoder": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/encoder/v1/encoder_grpc.pb.go",
+        "name": "EncoderServiceClient",
+        "methods": []
+    },
+    "gantry": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/gantry/v1/gantry_grpc.pb.go",
+        "name": "GantryServiceClient",
+        "methods": []
+    },
+    "generic_component": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/generic/v1/generic_grpc.pb.go",
+        "name": "GenericServiceClient",
+        "methods": []
+    },
+    "gripper": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/gripper/v1/gripper_grpc.pb.go",
+        "name": "GripperServiceClient",
+        "methods": []
+    },
+    "input_controller": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/inputcontroller/v1/input_controller_grpc.pb.go",
+        "name": "InputControllerServiceClient",
+        "methods": []
+    },
+    "motor": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/motor/v1/motor_grpc.pb.go",
+        "name": "MotorServiceClient",
+        "methods": []
+    },
+    "movement_sensor": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/movementsensor/v1/movementsensor_grpc.pb.go",
+        "name": "MovementSensorServiceClient",
+        "methods": []
+    },
+    "power_sensor": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/powersensor/v1/powersensor_grpc.pb.go",
+        "name": "PowerSensorServiceClient",
+        "methods": []
+    },
+    "sensor": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/sensor/v1/sensor_grpc.pb.go",
+        "name": "SensorServiceClient",
+        "methods": []
+    },
+    "servo": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/component/servo/v1/servo_grpc.pb.go",
+        "name": "ServoServiceClient",
+        "methods": []
+    },
+    "data_manager": {
+        "url": "https://github.com/viamrobotics/api/blob/main/service/datamanager/v1/data_manager_grpc.pb.go",
+        "name": "DataManagerServiceClient",
+        "methods": []
+    },
+    "generic_service": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/generic/v1/generic_grpc.pb.go",
+        "name": "GenericServiceClient",
+        "methods": []
+    },
+    "mlmodel": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/mlmodel/v1/mlmodel_grpc.pb.go",
+        "name": "MLModelServiceClient",
+        "methods": []
+    },
+    "motion": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/motion/v1/motion_grpc.pb.go",
+        "name": "MotionServiceClient",
+        "methods": []
+    },
+    "navigation": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/navigation/v1/navigation_grpc.pb.go",
+        "name": "NavigationServiceClient",
+        "methods": []
+    },
+    "slam": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/slam/v1/slam_grpc.pb.go",
+        "name": "SLAMServiceClient",
+        "methods": []
+    },
+    "vision": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/service/vision/v1/vision_grpc.pb.go",
+        "name": "VisionServiceClient",
+        "methods": []
+    },
+    "app": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/app/v1/app_grpc.pb.go",
+        "name": "AppServiceClient",
+        "methods": []
+    },
+    "billing": {
+        "url": "https://github.com/viamrobotics/api/blob/main/app/v1/billing_grpc.pb.go",
+        "name": "BillingServiceClient",
+        "methods": []
+    },
+    "data": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/app/data/v1/data_grpc.pb.go",
+        "name": "DataServiceClient",
+        "methods": []
+    },
+    "dataset": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/app/dataset/v1/dataset_grpc.pb.go",
+        "name": "DatasetServiceClient",
+        "methods": []
+    },
+    "data_sync": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/app/datasync/v1/data_sync_grpc.pb.go",
+        "name": "DataSyncServiceClient",
+        "methods": []
+    },
+    "robot": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/robot/v1/robot_grpc.pb.go",
+        "name": "RobotServiceClient",
+        "methods": []
+    },
+    "mltraining": {
+        "url": "https://raw.githubusercontent.com/viamrobotics/api/main/app/mltraining/v1/ml_training_grpc.pb.go",
+        "name": "MLTrainingServiceClient",
+        "methods": []
+    }
+}
